+++ WindowsCredentialUtility.sln
@@ -3,24 +3,34 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.9.34701.34
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WindowsCredentialUtility", "Keeper.WindowsCredentialManager\WindowsCredentialUtility.csproj", "{22133531-F154-43CD-8894-774EC80C121A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WindowsCredentialUtility", "WindowsCredentialUtility\WindowsCredentialUtility.csproj", "{22133531-F154-43CD-8894-774EC80C121A}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WindowsCredentialUtilityTests", "Keeper.WindowsCredentialManager.Tests\WindowsCredentialUtilityTests.csproj", "{14601A25-0AC4-4AF9-BE02-3496D4BBD287}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WindowsCredentialUtilityTests", "WindowsCredentialUtilityTests\WindowsCredentialUtilityTests.csproj", "{14601A25-0AC4-4AF9-BE02-3496D4BBD287}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
 		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{22133531-F154-43CD-8894-774EC80C121A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{22133531-F154-43CD-8894-774EC80C121A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{22133531-F154-43CD-8894-774EC80C121A}.Debug|x64.ActiveCfg = Debug|x64
+		{22133531-F154-43CD-8894-774EC80C121A}.Debug|x64.Build.0 = Debug|x64
 		{22133531-F154-43CD-8894-774EC80C121A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{22133531-F154-43CD-8894-774EC80C121A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{22133531-F154-43CD-8894-774EC80C121A}.Release|x64.ActiveCfg = Release|x64
+		{22133531-F154-43CD-8894-774EC80C121A}.Release|x64.Build.0 = Release|x64
 		{14601A25-0AC4-4AF9-BE02-3496D4BBD287}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{14601A25-0AC4-4AF9-BE02-3496D4BBD287}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{14601A25-0AC4-4AF9-BE02-3496D4BBD287}.Debug|x64.ActiveCfg = Debug|x64
+		{14601A25-0AC4-4AF9-BE02-3496D4BBD287}.Debug|x64.Build.0 = Debug|x64
 		{14601A25-0AC4-4AF9-BE02-3496D4BBD287}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{14601A25-0AC4-4AF9-BE02-3496D4BBD287}.Release|Any CPU.Build.0 = Release|Any CPU
+		{14601A25-0AC4-4AF9-BE02-3496D4BBD287}.Release|x64.ActiveCfg = Release|x64
+		{14601A25-0AC4-4AF9-BE02-3496D4BBD287}.Release|x64.Build.0 = Release|x64
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
