+++ WindowsCredentialUtilityTests/WindowsCredentialUtilityTests.csproj
@@ -7,6 +7,7 @@
 
     <IsPackable>false</IsPackable>
     <IsTestProject>true</IsTestProject>
+    <Platforms>AnyCPU;x64</Platforms>
   </PropertyGroup>
 
   <ItemGroup>
@@ -21,7 +22,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Keeper.WindowsCredentialManager\Keeper.WindowsCredentialManager.csproj" />
+    <ProjectReference Include="..\WindowsCredentialUtility\WindowsCredentialUtility.csproj" />
   </ItemGroup>
 
   <ItemGroup>
